move dabatase methods to jsonHandler
